Loop detector extracted to a separate file
